feat(admin,common): 新增学院班级管理模块并优化档案查询
【变更内容】
- 新增SysCollege、SysClass实体类
- 新增CollegeService、ClassService接口及实现类
- 新增CollegeMapper、ClassMapper接口及XML映射
- 新增学院、班级管理控制器接口
- 重构ArchivePageQueryDTO：将学院字段改为collegeId
- 更新MhArchive实体：替换学院班级字段为ID类型
- 修复ArchiveMapper：新增并实现update方法
- 调整WarnMapper和WarnService：修复详情查询方法参数和返回类型
- 优化WarnMapper.xml：移除多余空格
- 更新README：新增基础接口文档并修正表数量统计
- 更新ArchiveMapper.xml：适配新的档案查询和更新逻辑

【关联信息】
- 相关文件: 所有新增/修改的实体、服务、控制器、XML映射文件
- 解决问题: 优化档案模块的学院班级关联查询，新增基础数据管理接口

+++ mhp-pojo/src/main/java/com/mhp/vo/WarnPageQueryVO.java
@@ -8,13 +8,13 @@ public class WarnPageQueryVO {
 
     // 预警事件字段
     private Long eventId;
-    private Integer studentId;
+    private Long studentId; 
     private Long assessmentId;
     private Integer eventLevel;
     private Integer eventStatus;
     private String triggerSource;
     private String createTime;
-    
+
     //学号
     private String studentNo;
     // 学生姓名
